Added test for invalid command, updated utility function
Added a test for an invalid command as FREE CHILD may not have additional elements outside of the handle response per IBM's documentation. Expanded on the utility function as well due to needing to handle the case where the `TerminalNode` variables are actually `List<TerminalNode>`. This is a result of using the `*` due to null variables as well.